refactor/PRI-91 Login crean up
