+++ vex_app/frontend/src/App.tsx
@@ -1,3 +1,4 @@
+// react hooks
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
